fix: separate method for etching
