+++ examples/deploy.rs
@@ -129,8 +129,6 @@ async fn main() -> anyhow::Result<()> {
             },
             recipient_address: sender_address,
             redeem_script: commit_tx.redeem_script,
-            #[cfg(feature = "rune")]
-            runestone: None,
         })
         .await?;
     debug!("reveal transaction: {reveal_transaction:?}");
