+++ examples/mint.rs
@@ -124,8 +124,6 @@ async fn main() -> anyhow::Result<()> {
             },
             recipient_address: sender_address,
             redeem_script: commit_tx.redeem_script,
-            #[cfg(feature = "rune")]
-            runestone: None,
         })
         .await?;
     debug!("reveal transaction: {reveal_transaction:?}");
