+++ examples/transfer.rs
@@ -123,8 +123,6 @@ async fn main() -> anyhow::Result<()> {
             },
             recipient_address: sender_address, // NOTE: it's correct, see README.md to read about how transfer works
             redeem_script: commit_tx.redeem_script,
-            #[cfg(feature = "rune")]
-            runestone: None,
         })
         .await?;
     debug!("reveal transaction: {reveal_transaction:?}");
